requirements.txt
streamlit>=1.32.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
pypdf>=4.0.0
huggingface-hub>=0.20.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
groq>=0.9.0
python-dotenv>=1.0.0
pandas>=1.4.0
